// apps/console/src/__generated__/core/LinkScenarioDialogFragment.graphql.ts
export type LinkScenarioDialogFragment$data = {
  readonly id: string;
  readonly riskAssessmentScenarios: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly description: string | null | undefined;
        readonly id: string;
        readonly name: string;
      };
    }>;
  } | null | undefined;
  readonly " $fragmentType": "LinkScenarioDialogFragment";
};
export type LinkScenarioDialogFragment$key = {
  readonly " $data"?: LinkScenarioDialogFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"LinkScenarioDialogFragment">;
};